refactor: move car package
