+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/domain/car/CarPosition.kt
@@ -1,4 +1,4 @@
-package org.gonza.kotlinplayground.domain;
+package org.gonza.kotlinplayground.domain.car;
 
 class CarPosition(
     var value: Int = 0
